requirements.txt
streamlit==0.57.3
numpy==1.18.2
Pillow==7.2.0
tensorflow==2.2.0
